+++ db/migrate/20170126165413_add_reviewed_to_hosting_requests.rb
@@ -1,0 +1,6 @@
+class AddReviewedToHostingRequests < ActiveRecord::Migration[5.0]
+  def change
+    add_column :hosting_requests, :reviewed, :boolean, null: false, default: false
+    add_index :hosting_requests, :reviewed
+  end
+end